+++ src/main/java/disaster/pulse/api/repository/EventoRepository.java
@@ -3,4 +3,4 @@ package disaster.pulse.api.repository;
 import disaster.pulse.api.model.Evento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventoRepository extends JpaRepository<Evento, Integer> { }
+public interface EventoRepository extends JpaRepository<Evento, Long> { }
